requirements.txt
altgraph
asgiref
beautifulsoup4
blinker
certifi
charset-normalizer
click
crispy-bootstrap5
distlib
dj-database-url
Django
django-crispy-forms
django-email
django-environ
django-filter
djangorestframework
filelock
Flask
idna
itsdangerous
Jinja2
joblib
Markdown
MarkupSafe
numpy
opencv-contrib-python
packaging
pefile
pillow
pipenv
platformdirs
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
pyinstaller
pyinstaller-hooks-contrib
Pympler
PyMsgBox
PyPDF2
pyperclip
PyRect
PyScreeze
pyttsx3
pytweening
pywhatkit
qrcode
requests
scikit-learn
scipy
setuptools
soupsieve
sqlparse
threadpoolctl
typing_extensions
tzdata
urllib3
virtualenv
Werkzeug
wikipedia
